Split downloads and library screens into components
